Added a new parameter to heatmapHIV: filter
If the value of this parameter is not NULL, the rows of the matrix will
be ordered using the mean values of the columns which names match filter
value, in decreasing order. The columns of the matrix will also be
ordered to show the selected columns first in the heatmap. If filter
value is NULL, columns will be clustered using hclust and rows won't be
reordered.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>